Minor improvements/fixes
- Get counts for the 'No Context' context
- Recalculate the 'remaining' count with math
- Add 'isComplete()', 'isRemaining()', etc. functions to Task
- Fix filtering logic
- Rename blockedByDefer => deferred
- Pass down 'blocked' attribute (if parent is blocked, so is child)

+++ app/src/main/java/nu/huw/clarity/ui/viewholder/ContextViewHolder.java
@@ -35,9 +35,9 @@ public class ContextViewHolder extends ListAdapter.ViewHolder {
     public void bind(Context context, android.content.Context androidContext) {
 
         this.entry = context;
-        int available = this.entry.countAvailable;
-        int dueSoon   = this.entry.countDueSoon;
-        int overdue   = this.entry.countOverdue;
+      long available = this.entry.countAvailable;
+      long dueSoon = this.entry.countDueSoon;
+      long overdue = this.entry.countOverdue;
 
         Resources res = androidContext.getResources();
